requirements.txt
langgraph>=0.2.59
langchain>=0.3.20
langchain-core>=0.3.41
langsmith>=0.2.11
pydantic>=2.10.5
google-genai>=1.0.0
openpyxl>=3.1.5
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
